# Conflicts:
#	packages/jsii/lib/validator.ts
#	yarn.lock

+++ packages/jsii/test/deprecation-warnings.test.ts
@@ -78,8 +78,8 @@ function testpkg_Baz(p) {
         export interface Foo {}
         export interface Bar {}
         export interface Baz {
-          readonly foo: Foo; 
-          readonly bar: Bar; 
+          readonly foo: Foo;
+          readonly bar: Bar;
           readonly x: string;
         }
         `,
@@ -166,32 +166,13 @@ function testpkg_Baz(p) {
     );
   });
 
-  test('generates a call to print if the type is deprecated', async () => {
-    const result = await compileJsiiForTest(
-      `
-        ${DEPRECATED}
-        export interface Foo {}
-        `,
-      undefined /* callback */,
-      { addDeprecationWarnings: true },
-    );
-
-    expect(jsFile(result, '.warnings.jsii')).toMatch(`function testpkg_Foo(p) {
-    if (p == null)
-        return;
-    visitedObjects.add(p);
-    print("testpkg.Foo", "Use something else");
-    visitedObjects.delete(p);
-}`);
-  });
-
   test('generates functions for enums', async () => {
     const result = await compileJsiiForTest(
       `
         export enum State {
           ON,
-          
-          ${DEPRECATED} 
+
+          ${DEPRECATED}
           OFF
         }
         `,
@@ -205,6 +186,8 @@ function testpkg_Baz(p) {
         return;
     visitedObjects.add(p);
     const ns = require("./index.js");
+    if (Object.values(ns.State).filter(x => x === p).length > 1)
+        return;
     if (p === ns.State.OFF)
         print("testpkg.State#OFF", "Use something else");
     visitedObjects.delete(p);
@@ -212,37 +195,88 @@ function testpkg_Baz(p) {
 `);
   });
 
-  test('generates calls for supertypes', async () => {
+  test('generates calls for deprecated inherited properties', async () => {
     const result = await compileJsiiForTest(
       `
-        export interface Foo {}
-        export interface Bar {readonly foo: Foo;}
-        export interface Baz extends Bar {readonly x: string;}
+        export interface Baz {
+          /** @deprecated message from Baz */
+          readonly x: string;          
+        }
+        export interface Bar {
+          /** @deprecated message from Bar */
+          readonly x: string;          
+        }
+        export interface Foo extends Bar, Baz {
+        }
         `,
       undefined /* callback */,
       { addDeprecationWarnings: true },
     );
 
-    expect(jsFile(result, '.warnings.jsii')).toMatch(`function testpkg_Baz(p) {
+    const warningsFileContent = jsFile(result, '.warnings.jsii');
+
+    // For each supertype, its corresponding function should be generated, as usual
+    expect(warningsFileContent).toMatch(`function testpkg_Baz(p) {
     if (p == null)
         return;
     visitedObjects.add(p);
-    testpkg_Bar(p);
+    if ("x" in p)
+        print("testpkg.Baz#x", "message from Baz");
     visitedObjects.delete(p);
-}
-`);
+}`);
+    expect(warningsFileContent).toMatch(`function testpkg_Bar(p) {
+    if (p == null)
+        return;
+    visitedObjects.add(p);
+    if ("x" in p)
+        print("testpkg.Bar#x", "message from Bar");
+    visitedObjects.delete(p);
+}`);
+
+    // But a call for one of the instances of the property should also be generated in the base function
+    expect(warningsFileContent).toMatch(`function testpkg_Foo(p) {
+    if (p == null)
+        return;
+    visitedObjects.add(p);
+    if ("x" in p)
+        print("testpkg.Baz#x", "message from Baz");
+    visitedObjects.delete(p);
+}`);
   });
 
-  test('generates calls for tyes with deprecated properties', async () => {
+  test('skips properties that are deprecated in one supertype but not the other', async () => {
+    const result = await compileJsiiForTest(
+      `
+        export interface Baz {
+          readonly x: string;          
+        }
+        export interface Bar {
+          /** @deprecated message from Bar */
+          readonly x: string;          
+        }
+        export interface Foo extends Bar, Baz {
+        }
+        `,
+      undefined /* callback */,
+      { addDeprecationWarnings: true },
+    );
+
+    const warningsFileContent = jsFile(result, '.warnings.jsii');
+
+    expect(warningsFileContent).toMatch(`function testpkg_Foo(p) {
+}`);
+  });
+
+  test('generates calls for types with deprecated properties', async () => {
     const result = await compileJsiiForTest(
       `
       export interface Bar {
         readonly x: string;
       }
-      
+
       export interface Foo {
         readonly y: string;
-    
+
         /** @deprecated kkkkkkkk */
         readonly bar: Bar;
       }
@@ -264,10 +298,40 @@ function testpkg_Baz(p) {
 `);
   });
 
+  test('generates calls for each property of a deprecated type', async () => {
+    const result = await compileJsiiForTest(
+      `
+      /** @deprecated use Bar instead */
+      export interface Foo {
+        readonly bar: string;
+        readonly baz: number;
+      }
+      `,
+      undefined /* callback */,
+      { addDeprecationWarnings: true },
+    );
+
+    expect(jsFile(result, '.warnings.jsii')).toMatch(`function testpkg_Foo(p) {
+    if (p == null)
+        return;
+    visitedObjects.add(p);
+    if ("bar" in p)
+        print("testpkg.Foo#bar", "use Bar instead");
+    if ("baz" in p)
+        print("testpkg.Foo#baz", "use Bar instead");
+    visitedObjects.delete(p);
+}
+`);
+  });
+
   test('generates calls for types in other assemblies', async () => {
+    const calcBaseOfBaseRoot = resolveModuleDir(
+      '@scope/jsii-calc-base-of-base',
+    );
     const calcBaseRoot = resolveModuleDir('@scope/jsii-calc-base');
     const calcLibRoot = resolveModuleDir('@scope/jsii-calc-lib');
 
+    await compile(calcBaseOfBaseRoot, false);
     await compile(calcBaseRoot, true);
     await compile(calcLibRoot, true);
     const warningsFile = loadWarningsFile(calcBaseRoot);
@@ -281,11 +345,11 @@ function testpkg_Baz(p) {
     // Recompiling without deprecation warning to leave the packages in a clean state
     await compile(calcBaseRoot, false);
     await compile(calcLibRoot, false);
-  }, 25000);
+  }, 30000);
 });
 
 describe('Call injections', () => {
-  test('does not add warnings, by default', async () => {
+  test('does not add warnings by default', async () => {
     const result = await compileJsiiForTest(
       `
     export class Foo {
@@ -316,15 +380,32 @@ describe('Call injections', () => {
       { addDeprecationWarnings: true },
     );
 
-    const expectedPath = ['..', '..', '.warnings.jsii.js'].join(path.sep);
-    const requireRegex = /const jsiiDeprecationWarnings = require\("(.+)"\);/g;
+    const expectedPath = ['..', '..', '.warnings.jsii.js'].join('/');
 
     const content = jsFile(result, 'some/folder/source');
-    const match = requireRegex.exec(content);
-    expect(match).toBeDefined();
+    expect(content).toContain(
+      `const jsiiDeprecationWarnings = require("${expectedPath}")`,
+    );
+  });
 
-    const actualPath = match![1];
-    expect(path.normalize(actualPath)).toEqual(expectedPath);
+  test('does not generate a require statement when no calls were injected', async () => {
+    const result = await compileJsiiForTest(
+      {
+        'index.ts': `export * from './some/folder/handler'`,
+        'some/folder/handler.ts': `
+          export function handler(event: any) { return event; }
+        `,
+      },
+      undefined /* callback */,
+      { addDeprecationWarnings: true },
+    );
+
+    const expectedPath = ['..', '..', '.warnings.jsii.js'].join('/');
+
+    const content = jsFile(result, 'some/folder/handler');
+    expect(content).not.toContain(
+      `const jsiiDeprecationWarnings = require("${expectedPath}")`,
+    );
   });
 
   test('deprecated methods', async () => {
@@ -385,7 +466,7 @@ describe('Call injections', () => {
     export class Foo {
       private _x = 0;
       public get x(){return this._x}
-      
+
       ${DEPRECATED}
       public set x(_x: number) {this._x = _x;}
     }
@@ -412,7 +493,7 @@ describe('Call injections', () => {
     );
 
     expect(jsFile(result)).toMatch(
-      'constructor() { jsiiDeprecationWarnings.print("testpkg.Foo", ""); }',
+      'constructor() { jsiiDeprecationWarnings.print("testpkg.Foo", "Use something else"); }',
     );
   });
 });
